+++ dantro/plot_creators/ext_funcs/mpl_setup.py
@@ -2,8 +2,7 @@
 
 import logging
 
-import matplotlib
-matplotlib.use("Agg")
+import matplotlib as mpl
 import matplotlib.pyplot as plt